create new table Images and related with posts

+++ Repo_Core/Services/IBlogService.cs
@@ -4,7 +4,7 @@ namespace Repo_Core.Services
 {
     public interface IBlogService
     {
-        ICollection<Posts> GetPosts(int page, int pagesize);
+        ICollection<Posts> GetPosts(int page, byte pagesize);
         Task<Posts> Add(Posts blogs);
     }
 }
